// web/platform-admin-web/src/api/generated/model/getMemberPageFiltersDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */
import type { GetMemberPageFiltersDtoStatus } from './getMemberPageFiltersDtoStatus';

export interface GetMemberPageFiltersDto {
  /** 멤버 상태 필터 */
  status?: GetMemberPageFiltersDtoStatus;
  /** 검색어 */
  search?: string;
}