// web/platform-admin-web/src/api/generated/model/memberPageItemStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 멤버 상태
 */
export type MemberPageItemStatus = typeof MemberPageItemStatus[keyof typeof MemberPageItemStatus];


export const MemberPageItemStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;