// web/platform-admin-web/src/api/generated/model/membersControllerGetMemberPageV1DirectionItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

export type MembersControllerGetMemberPageV1DirectionItem = typeof MembersControllerGetMemberPageV1DirectionItem[keyof typeof MembersControllerGetMemberPageV1DirectionItem];


export const MembersControllerGetMemberPageV1DirectionItem = {
  asc: 'asc',
  desc: 'desc',
} as const;