// web/platform-admin-web/src/api/generated/model/membersControllerGetMemberPageV1SortItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

export type MembersControllerGetMemberPageV1SortItem = typeof MembersControllerGetMemberPageV1SortItem[keyof typeof MembersControllerGetMemberPageV1SortItem];


export const MembersControllerGetMemberPageV1SortItem = {
  createdAt: 'createdAt',
  name: 'name',
  status: 'status',
  lastLoginAt: 'lastLoginAt',
} as const;